[ao] Implemented Outlier Detection Report Generation
Summary: This adds the implementation for the report generation for the
Outlier Detector class. This includes both the generation of a
dictionary containing each module that had an observer attached and any
relavent stats collected by the observer that can help shed light on
outlier relavent data or computed metrics. It also includes a string
denoting specific modules that had outliers and gives a bit of insight
into what channels they are contained in.

This contains both the implementation for the report generation for the
outlier detector as well as a test class to test the report generation
functionality.

Test Plan: python test/test_quantization.py TestFxDetectOutliers

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 7e4ff5a94ccbe2cae69e87dc944e919e8834fc41
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80937